requirements.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.15.2
django-jazzmin==3.0.1
dotenv==0.9.9
python-dotenv==1.2.1
redis==5.2.1
sqlparse==0.5.3
whitenoise==6.8.2
